Chunk 2: Finish decomposing the site presenter into section modules
- Extract alternatives section to sections/alternatives.ts
- Extract long-range section to sections/long-range.ts
- Extract forecast/days-ahead to sections/forecast.ts
- Extract footer to sections/footer.ts
- Extract session recommendation to sections/session-rec.ts
- Extract kit advisory to sections/kit-advisory.ts
- Extract creative spark to sections/creative-spark.ts
- Extract hourly outlook to sections/hourly-outlook.ts
- Extract spur of the moment to sections/spur-of-moment.ts
- Extract daylight utility bar to sections/daylight-utility.ts
- Create shared.ts for common site presenter helpers (sCard, sPill, etc.)
- Reduce format-site.ts from 743 lines to 282 lines
- All existing tests pass

Closes #109
